requirements.txt
requests
beautifulsoup4
openai
python-dotenv
flask
logtail-python
gunicorn 